Add a variable for `make.conf` to select collectors.
This updates the ebuild to look for a variable in make.conf called
`NETDATA_COLLECTORS` which contains a list of collectors to enable
support for in the build. This helps keep USE flags more sanely
manageable both in the ebuild and for users of the package.

Explicit compatability USE flags are kept for the few collectors that
already had USE flags (they will be removed at some point in the future,
as some of them are at-odds with standard USE flags used for other
purposes).

For the internal collectors this pulls in the required packages and
enables the required configure flags.

For the python collector modules, this ensures that python support is
enabled.

Dependency checks based on this will be added in the next few commits
for collectors which require additional dependencies.