Add depenndencies for Python collectors.
THis adds the hard dep on the `sudo` USE flag for collectors which need
it, as wella s pulling in required packages for collectors that have
requirements beyond what is normally included in an install of the
software they monitor on Gentoo.

The dockerd collector probably needs additional dependencies but I have
not been able to decipher how all the Docker packages in Portage are
inter-related to figure out what it actually needs to pull in.

The oracledb collector may require additional dependencies, but I do not
have access to the Oracle components that are required to determine this
properly.