Change the prefix on the collector USE flags.
The earlier commit claiming that we could use a variable in make.conf
forgot about the rules surrounding USE_EXPAND. Because we can't modifiy
that variable, we can't have a variable for the collector USE flags.

Due to this, this commit changes the flag names to be nicer to work
with, but still keeps a flag per-collector.