Update KEYWORDS to reflect actua support.
We functionally support 32/64 bit x86 and 32 bit ARM, though Netdata
should work just fine on most other platforms Gentoo supports.

This changes the KEYWORDS for the regular Netdata ebuilds so that it's
listed as stable on x86, amd64, and arm, and unstable on most other
arches (all the ones listed as stable except s390 have been build tested
by @Ferroin). alpha, hppa, ia64, and sh are excluded because I have no
way to verify that they work or even try to reproduce any issues that
come up.  the non-linux keywords are excluded because I don't have the
immediate ability to test them, but they may be added in the near future.

The net upshot of this is that users on x86, amd64, and arm will have an
even easier time installing Netdata.